Seperated Controllers for Admin site and User site
